requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
sentence-transformers
Pillow
pymupdf
pytesseract
numpy
openai
pytest
httpx